Add files via upload
Added possibility to choose period of forecast from index as well as date.
Modified the forecast date to take index of end-date +1 to avoid having a 1-timestamp forecast only.
Added Modes comparison between forecast and real data
Added Frobenius norm calculation of the difference between modes of forecast and mode of real data. That allows to visualize which areas are found well with the forecast, and adds a metric to it rather than just relying on the human eye.